Enforce max block number.
RFC6330 lets you have 256 blocks at most. Check at
the interleaver definition.

The C interface did not check for correct C++ initialization.
fix that, too.

+++ test/rfc_test.cpp
@@ -377,7 +377,7 @@ uint64_t decode (uint32_t mysize, std::mt19937_64 &rnd, float drop_prob,
 	if (micro1 == 0)
 		return 0;
 
-	if (drop_prob > 100.0)
+	if (drop_prob > static_cast<float>(100.0))
 		drop_prob = 90.0;	// this is still too high probably.
 	std::uniform_real_distribution<float> drop (0.0, 100.0);
